refactor(extractor): move OASF extractor and nlsearch from client to utils
