+++ utils/extractor/assets.go
