+++ utils/extractor/assets_test.go
